awstesting: Correct potential circlar reference with SDK test client
