+++ awstesting/client.go
@@ -5,7 +5,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/aws/client/metadata"
 	"github.com/aws/aws-sdk-go/aws/defaults"
-	"github.com/aws/aws-sdk-go/awstesting/mock"
 )
 
 // NewClient creates and initializes a generic service client for testing.
@@ -18,25 +17,4 @@ func NewClient(cfgs ...*aws.Config) *client.Client {
 	def.Config.MergeIn(cfgs...)
 
 	return client.New(*def.Config, info, def.Handlers)
-}
-
-// NewMockClient creates and initializes a client that will connect to the
-// mock server
-func NewMockClient(cfgs ...*aws.Config) *client.Client {
-	c := mock.Session.ClientConfig("Mock", cfgs...)
-
-	svc := client.New(
-		*c.Config,
-		metadata.ClientInfo{
-			ServiceName:   "Mock",
-			SigningRegion: c.SigningRegion,
-			Endpoint:      c.Endpoint,
-			APIVersion:    "2015-12-08",
-			JSONVersion:   "1.1",
-			TargetPrefix:  "MockServer",
-		},
-		c.Handlers,
-	)
-
-	return svc
 }
